+++ Sources/Blocked.Event.swift
@@ -3,7 +3,7 @@ import Foundation
 extension Blocked {
     public struct Event {
         let timestamp: UInt32
-        let domain: Int
-        let tracker: Int
+        let domain: UInt16
+        let tracker: UInt16
     }
 }
